use slug as id for fetching collectibles (#649)
* use slug as id for fetching collectibles

* fix tests

* review fixes

* fix

* fix

* fix tests

* fix tests

* merge
