+++ pkg/integration/testdata/coin_fixtures.json
@@ -69,7 +69,7 @@
     "address": "0x0875BCab22dE3d02402bc38aEe4104e1239374a7",
     "owner": "0x0875BCab22dE3d02402bc38aEe4104e1239374a7",
     "collections": "0x0875BCab22dE3d02402bc38aEe4104e1239374a7",
-    "collection_id": "0xfaafdc07907ff5120a76b34b731b278c38d6043c"
+    "collection_id": "ens"
   },
   "classic": {
     "address": "0x0875BCab22dE3d02402bc38aEe4104e1239374a7",
